+++ CHANGELOG.md
@@ -4,7 +4,9 @@ This repo attempts to follow [semantic versioning](https://semver.org/).
 
 ## Unreleased
 
-- None
+### Fixes
+
+- Fixed the handling of the `skipToDate` OccurrenceIterator `next()` arg ([!54](https://gitlab.com/john.carroll.p/rschedule/-/merge_requests/54))
 
 ## 1.0.4 FIX (2020/6/1)